Cache view query string since views get disposed
Views are tied to a handle that should be properly disposed. When this occurs, accessing the hash code - which is from the handle - throws an exception. This was causing the property cache to overflow. Instead, column information can be cached based on the unique query string which also improves performance for future lookups on the same data view.