Add additional properties to record objects
Added the PSStandardMembers MemberSet to each Record object to maintain that just the column properties are visible by default. The databases' PSPath and query string used to retrieve the records were added.

Had to add those properties to the Members collection; otherwise, adapted properties - cached in a static collection deep in PowerShell - would get copied as soon as the Properties collection property was accessed. This resulted in the cached values from previous records being copied as well.

To avoid unnecessary allocations, moved caching into the ViewManager. The ViewManager was moved into the PowerShell namespace since it now dealt specifically with PowerShell types, though it was always specific to PowerShell cmdlets anyway.