Add variable support to attribute display
Support for a new variable is defined to alter the display value of an attribute column. The value type is not changed, however, so any comparisons will still work. This does not export a variable from the module because of potential issues in PowerShell under investigation. So an about_MSI topic was added to describe this and any future option variables as well as brief information about this module.