Map component paths to correct registry keys
Fixes issue #14457 by creating a mapping of 32- and 64-bit registry keys based on MSDN documentation for the registry reidrector. This is a best effort for the current implementation of WOW64 and may not always be 100% correct, but without any APIs exported to do the mapping there's really no choice with any decent performance.